Remove legacy mysql wrapper and implement mysqli DB object
